// src/Models/ViewModels/StudentListViewModel.ts
class StudentListViewModel {

    id: number = 0;
    name: string = "";
    surname: string = "";
    birthday: string = "";
    nationality: string = "";
    specialityId: number = 0;
    status: string = "";
    sector: string = "";
    fatherName: string = "";
    motherName: string = "";
    specialityName: string = "";
    subjectName: string = "";
    minds: string = "";
    mindType: string = "";
    phoneMain: string = "";
    phoneWhatsApp: string = "";
    phoneFather: string = "";
    phoneMother: string = "";
    regionName: string = "";
    userOpen: string = "";
    register: string = "";
    callDate: string = ""
}

export { StudentListViewModel };